refactor: rename StructuredOutputScorer to follow camelCase convention
- Rename StructuredOutputScorer.ts to structuredOutputScorer.ts
- Update imports in search-events-agent.eval.ts and search-issues-agent.eval.ts
- Remove obsolete CLEANUP_PLAN.md
