+++ packages/mcp-server-evals/src/evals/utils/structuredOutputScorer.ts
